requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-filter==24.3
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.7.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-environ==0.11.2
django-cors-headers==4.4.0
dj-database-url==2.2.0
redis==5.0.8
sentry-sdk==2.20.0
gunicorn==22.0.0
prometheus-client==0.21.1

django-storages==1.14.3
boto3==1.34.162

django-guardian==2.4.0
django-otp==1.3.0
django-two-factor-auth==1.17.0
phonenumbers==8.13.40
xhtml2pdf==0.2.17
python-docx==1.1.2
opentelemetry-distro==0.52b0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-instrumentation-django==0.52b0
